Tx: get_ninputs and get_noutputs return usize

+++ src/transaction/transaction.rs
@@ -97,7 +97,7 @@ impl Transaction {
         };
 
         // In Counter - 1-9 tx_bytes
-        match buffer.write_varint(self.get_ninputs()) {
+        match buffer.write_varint(self.get_ninputs() as u64) {
             Ok(_) => (),
             Err(e) => return Err(BSVErrors::SerialiseTransaction("n_inputs".to_string(), e)),
         };
@@ -114,7 +114,7 @@ impl Transaction {
         }
 
         // Out Counter - 1-9 tx_bytes
-        match buffer.write_varint(self.get_noutputs()) {
+        match buffer.write_varint(self.get_noutputs() as u64) {
             Ok(_) => (),
             Err(e) => return Err(BSVErrors::SerialiseTransaction("n_outputs".to_string(), e)),
         };
@@ -176,13 +176,13 @@ impl Transaction {
     }
 
     #[wasm_bindgen(js_name = getInputsCount)]
-    pub fn get_ninputs(&self) -> u64 {
-        self.inputs.len() as u64
+    pub fn get_ninputs(&self) -> usize {
+        self.inputs.len()
     }
 
     #[wasm_bindgen(js_name = getOutputsCount)]
-    pub fn get_noutputs(&self) -> u64 {
-        self.outputs.len() as u64
+    pub fn get_noutputs(&self) -> usize {
+        self.outputs.len()
     }
 
     #[wasm_bindgen(js_name = getInput)]
